requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
langchain>=0.1.10,<0.2.0
langchain-community>=0.0.10,<0.1.0
google-generativeai==0.5.4
waitress==2.1.2